web/requirements.txt
Flask-Bootstrap4-4.0.2
Flask-WTF-0.14.3